+++ src/providers/BitComponentsProvider/Item.ts
@@ -1,10 +1,12 @@
+import * as vscode from "vscode"
 
-
-export default class Item {
+export default class Item extends vscode.TreeItem {
 
     constructor(
         public name: string,
         public children?: Item[]
-    ) { }
+    ) { 
+        super(name);
+    }
 
 }