+++ src/providers/BitComponentsProvider/Scope.ts
@@ -1,7 +1,9 @@
 import Item from "./Item";
-
+import * as vscode from "vscode";
 
 export default class Scope extends Item {
+
+    collapsibleState= vscode.TreeItemCollapsibleState.Expanded;
 
     constructor(
         public name: string,