+++ src/utils/bit/Component.ts
@@ -1,0 +1,16 @@
+
+export interface File{
+    relativePath: string;
+    test:boolean;
+    name: string;
+}
+export default interface Component{
+    files: File[];
+    mainFile: string;
+    trackDir: string;
+    rootDir: string;
+    origin: string;
+    exported: boolean;
+}
+
+export type NameAndComponent = [string, Component]